Now handling up/down drag events
In the initial Flickity fork used for this fork, drag events weren't being handled differently for vertical cells, meaning that a user still had to swipe left-to-right or right-to-left to cycle through cells. Discovered instances of drag and animation handling in three separate files (animate.js, flickity.js, and drag.js), and added conditionals testing for the verticalCells option, then handled positioning and dragging on the y-axis, a opposed to the x.